Move ContextInspectable to covia.adapter.agent
The interface only makes sense alongside the LLM-backed transition
adapters that implement it (LLMAgentAdapter, GoalTreeAdapter), all of
which live in covia.adapter.agent. Co-locating it makes the inheritance
graph local to the package and removes a needless cross-package import
from those adapters.

The single non-agent caller (AgentAdapter.handleContext, which
dispatches via the interface) gains an explicit import — fine, it's
the one boundary crossing.
