+++ venue/src/main/java/covia/adapter/AgentAdapter.java
@@ -25,6 +25,7 @@ import convex.core.data.prim.CVMBool;
 import convex.core.data.prim.CVMLong;
 import convex.core.lang.RT;
 import convex.core.util.Utils;
+import covia.adapter.agent.ContextInspectable;
 import covia.api.Fields;
 import covia.grid.Job;
 import covia.grid.Status;
