+++ venue/src/main/java/covia/adapter/agent/ContextInspectable.java
@@ -1,4 +1,4 @@
-package covia.adapter;
+package covia.adapter.agent;
 
 import convex.core.data.ACell;
 import convex.core.data.AMap;
